Fix connection deprecation warning
Again the Quick start is omitting stuff that you actually need. This time it's a deprecation warning that a redis connection, either as an instance of [IORedis](https://github.com/luin/ioredis/blob/master/API.md) or as config, must be provided when instantiating Queues, Workers, and QueueEvents.

> BullMQ: DEPRECATION WARNING! Optional instantiation of Queue, Worker, QueueScheduler and QueueEvents without providing explicitly a connection or connection options is deprecated. This behaviour will be removed in the next major release

It _currently_ works without it, but will break in future versions. So, to remove the warning and ensure the project runs in the future we specify connection details.

And because your redis might be different to mine, we do it in such a way that it can actually be configured.