Move call to addJobs() to end of file
Each run gives slightly different output with not all messages being seen. It just seems to be a timing thing. But you would expect all the setup to have happened (queues, workers etc) before the jobs are created and added to the queues.

So, we move the call to addJobs() to the end of the file.