Clear out our jobs by default
If you explore Redis using a GUI client you'll see a collection that matches your queue name and an ever expanding list of hashes that match the jobs.

This is because by default BullMQ doesn't delete jobs once they've been processed. So, to avoid this getting out of hand you can tell BullMQ to delete the job data once it's complete or failed.

We do this on the Queue so any jobs added to it will have this config applied automatically. But for reference you can set this config when adding the job. So, within the same queue you can have jobs that persists and ones that don't should you need to.